% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_knit_jobs.R
\name{.__dot_to_param_str__}
\alias{.__dot_to_param_str__}
\title{Helper function to convert ellipsis arguments to string to add to glue}
\usage{
.__dot_to_param_str__(..., comma_prefix = FALSE)
}
\arguments{
\item{...}{Ellipsis arguments passed from parent call.}

\item{comma_prefix}{Whether to add a comma before}
}
\value{
A string
}
\description{
Helper function to convert ellipsis arguments to string to add to glue
}
