% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_knit_jobs.R
\name{.__interactive_knit_jobs__}
\alias{.__interactive_knit_jobs__}
\title{Helper function to create knitting script and call RStudio job to run that script.}
\usage{
.__interactive_knit_jobs__(file, ...)
}
\arguments{
\item{file}{Rmd file to knit}

\item{...}{additional parameters passed to rmarkdown::render}
}
\value{
A string containing job's ID.
}
\description{
Helper function to create knitting script and call RStudio job to run that script.
}
