% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_name.R
\name{get_new_name_quarto}
\alias{get_new_name_quarto}
\title{Function to get new file names, Quarto version}
\usage{
get_new_name_quarto(v, orig_name, orig_dir, folders)
}
\arguments{
\item{v}{Version list}

\item{orig_name}{Original file name}

\item{orig_dir}{Original file directory}

\item{folders}{Folders to store the new generated files}
}
\value{
New file name
}
\description{
Function to get new file names, Quarto version
}
