% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution_chunks.R
\name{get_solution_chunks}
\alias{get_solution_chunks}
\title{Function to label the solution chunks correctly}
\usage{
get_solution_chunks(version_chunk, all_info)
}
\arguments{
\item{version_chunk}{Chunks to knit in the documents}

\item{all_info}{All the code chunks in the data}
}
\value{
Returns all the code chunks with the solutions correctly labeled as so
}
\description{
Function to label the solution chunks correctly
}
