% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_chunks.R
\name{get_version_chunks}
\alias{get_version_chunks}
\title{Gets version tag information from chunks
This function extracts the chunks ```
Helper for \code{version()}}
\usage{
get_version_chunks(source_text)
}
\arguments{
\item{source_text}{input text}
}
\value{
Version tags from chunks
}
\description{
Gets version tag information from chunks
This function extracts the chunks ```
Helper for \code{version()}
}
