% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version_text.R
\name{get_version_text}
\alias{get_version_text}
\title{Gets version tag information from text
This function extracts the sections \%\%\%
Helper for \code{version()}}
\usage{
get_version_text(source_text)
}
\arguments{
\item{source_text}{input text}
}
\value{
Versioned sections
}
\description{
Gets version tag information from text
This function extracts the sections \%\%\%
Helper for \code{version()}
}
