% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_versions_opts.R
\name{knit_versions_opt}
\alias{knit_versions_opt}
\title{Generator of knit_versions.}
\usage{
knit_versions_opt(...)
}
\arguments{
\item{...}{Whatever you want to edit the knit_versions() function. See \link{knit_versions}}
}
\value{
A partial call of knit_versions.
}
\description{
This function is intended to be used in the context of Rmd files only
}
