% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_orig_text.R
\name{prep_orig_text}
\alias{prep_orig_text}
\title{Function to eliminate all calls to \code{versions()} and setup chunks
#' Helper for \code{version()}}
\usage{
prep_orig_text(orig_text, global_eval = TRUE)
}
\arguments{
\item{orig_text}{Text from the original RMd}

\item{global_eval}{If to execute the code chunks in the document or not}
}
\value{
Pre-processed text containing a cleaned-up RMd
}
\description{
Function to eliminate all calls to \code{versions()} and setup chunks
#' Helper for \code{version()}
}
