% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_cleaner.R
\name{version_cleaner}
\alias{version_cleaner}
\title{Function to clean the different versions.}
\usage{
version_cleaner(v, orig_text, sec_info, all_info, orig_dir)
}
\arguments{
\item{v}{Version list}

\item{orig_text}{Original file text}

\item{sec_info}{Sections info}

\item{all_info}{All chunks info}

\item{orig_dir}{Original file directory}
}
\value{
Versioned cleaned text
}
\description{
Function to clean the different versions.
}
