% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{versions}
\alias{versions}
\title{Knit different versions of a file from chunk and section options}
\usage{
versions(global_eval = TRUE, to_knit = NULL, folders = NULL)
}
\arguments{
\item{global_eval}{Logical.}

\item{to_knit}{Character vector specifying which versions to write and knit
into separate files.  If not specified, all versions are produced.}

\item{folders}{List of versions and subfolder to put them in. Use pattern
\code{version_name = folder_name}.  Default is each version in its own folder.}
}
\value{
none
}
\description{
\code{versions} is a function that should be included in the setup chunk of
an R Markdown document.  Its purpose is to write, then knit, R Markdown source
files for several versions of a document, such as different exams in a course.
}
