% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yaml_version.R
\name{write_yaml_version}
\alias{write_yaml_version}
\title{This function write new yaml to the versioned header}
\usage{
write_yaml_version(orig_text, version)
}
\description{
This function write new yaml to the versioned header
}
