% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_versions.R
\name{knit_versions}
\alias{knit_versions}
\title{Knit different versions of a file from chunk and section options}
\usage{
knit_versions(
  orig_file,
  global_eval = TRUE,
  to_knit = NULL,
  folders = NULL,
  knit_global = TRUE
)
}
\arguments{
\item{orig_file}{The original file serving as the template}

\item{global_eval}{Logical.}

\item{to_knit}{Character vector specifying which versions to write and knit
into separate files.  If not specified, all versions are produced.}

\item{folders}{List of versions and subfolder to put them in. Use pattern
\code{version_name = folder_name}.  Default is each version in its own folder.}

\item{knit_global}{Whether to knit the global file, default to TRUE}
}
\value{
none
}
\description{
\code{knit_versions} is a function that provides an independent version or \code{versions}
with extra options to give more authorities to the users.
}
