% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{versions}
\alias{versions}
\title{Knit different versions of a file from chunk and section options}
\usage{
versions(
  pull_solutions = TRUE,
  global_eval = TRUE,
  to_knit = NULL,
  folders = NULL
)
}
\arguments{
\item{pull_solutions}{Logical. Should we create separate solution files for
each version of the document?}

\item{global_eval}{Logical. Should we replaces the global \code{eval = FALSE} option
with \code{TRUE} in the individual versions?}

\item{to_knit}{Character vector specifying which versions to write and knit
into separate files.  If not specified, all versions are produced.}

\item{folders}{List of versions and subfolder to put them in. Use pattern
\code{version_name = folder_name}.  Default is each version in its own folder.}
}
\value{
none
}
\description{
\code{versions} is a function that should be included in the setup chunk of
an R Markdown document.  Its purpose is to write, then knit, R Markdown source
files for several versions of a document, such as different exams in a course.
}
\details{
Code chunks may be tagged as version-specific using the option \code{version}.
Text sections may also be tagged as version-specific using \code{\%\%\%} wrappers.
See example Rmd source below.

The version label \code{"solution"} is special if \code{pull_solutions = TRUE},
as it is combined with the other version labels to make a solution set.

The version label \code{"none"} is special; it will be ignored in the creation
of the child documents.  Use it to leave yourself notes in the original document.

Example R Markdown source:

\preformatted{
---
title: "Example"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
templar::versions()
```

\%\%\%
version: A

You are taking **Exam A**
\%\%\%

\%\%\%
version: B

You are taking **Exam B**
\%\%\%

## Question 1: Means

Find the mean of the vector `a`

```{r, version = "A"}
set.seed(123)
```

```{r, version = "B"}
set.seed(456)
```

```{r}
a <- rnorm(10)
```

\%\%\%
version: solution

The mean is `r mean(a)`
\%\%\%

\%\%\%
version: none

Note to self: make a version C later.
\%\%\%

}
}
